dodanie metody szukającej wycieczek po standardzie hotelu

+++ src/main/java/com/example/travelagency2/service/TravelService.java
@@ -60,7 +60,7 @@ public class TravelService {
         return travelRepository.findAllByHotelTypeIs(type);
     }
 
-    public List<TravelModel> findTravelsByHotelStars(Integer hotelStars) {
+    public List<TravelModel> findTravelsByHotelStars(Byte hotelStars) {
         return travelRepository.findAllByHotelModel_StarsNumber(hotelStars);
     }
